Added H3

+++ Homework/H3/H3.ino
@@ -98,7 +98,7 @@ void updateDisplay() {
     displayBlinkOn = !displayBlinkOn;
   }
  
-  for (size_t i = 0; i < 8; i++) {
+  for (int i = 0; i < 8; i++) {
     if (i == displayState) {
       digitalWrite(displayPins[i], displayBlinkOn ? HIGH : LOW);
     } else {
